refactor: service 분리 - 체스 방을 관리하는 서비스 - 체스판을 관리하는 서비스
